Add support to build with Visual Studio
This adds support to build libsoup and libsoup-gnome with Visual Studio
2008 through 2015 using project files.  This also build the example
programs, that can be used to test and demonstrate the use of libsoup.

Many thanks to Ignacio Casal Quinteiro for doing the work to update the
code to export the symbols and variables using compiler directives, and
also provide a reference for this set of project files from his github
repo.